split term to term and view
