+++ examples/custom_bindings/src/main.rs
@@ -56,12 +56,10 @@ impl Application for App {
                     monospaced: false,
                     stretch: Stretch::Normal,
                 },
-                ..iced_term::FontSettings::default()
             },
             theme: iced_term::ColorPalette::default(),
             backend: iced_term::BackendSettings {
                 shell: system_shell.to_string(),
-                ..iced_term::BackendSettings::default()
             },
         };
 
@@ -140,7 +138,7 @@ impl Application for App {
                             ),
                         );
                     },
-                    iced_term::Event::Ignored(_) => {},
+                    _ => {},
                 };
 
                 Command::none()
