+++ src/lib.rs
@@ -3,11 +3,11 @@ pub mod bindings;
 mod font;
 mod term;
 mod theme;
+mod view;
 
 pub use alacritty_terminal::term::TermMode;
-pub use backend::BackendSettings;
+pub use backend::settings::BackendSettings;
 pub use font::FontSettings;
-pub use term::{
-    term_view, Command, Event, Term, TermSettings, TermView, TermViewState,
-};
+pub use term::{Command, Event, Term, TermSettings};
 pub use theme::{ColorPalette, TermTheme};
+pub use view::{term_view, TermView, TermViewState};
